vendor: Import r3583 from http://svn.netlabs.org/repos/kbuild/trunk.
    Directories removed when importing:
    - /dist
    - /kBuild/bin
    - /SlickEdit

    Note that this also brings /src/lib/kStuff which is an external
    repository of its own (http://trac.netlabs.org/kstuff/) but we will
    keep it here instead of making a git submodule as there is no
    kstuff git repo yet.

+++ src/lib/nt/nthlpcore.c
@@ -1,4 +1,4 @@
-/* $Id: nthlpcore.c 2998 2016-11-05 19:37:35Z bird $ */
+/* $Id: nthlpcore.c 3534 2021-12-20 23:31:55Z bird $ */
 /** @file
  * MSC + NT core helpers functions and globals.
  */
@@ -118,7 +118,7 @@ void birdResolveImportsWorker(void)
                 || !WriteFile(GetStdHandle(STD_ERROR_HANDLE), pszSym, (DWORD)strlen(pszSym), &cbWritten, NULL)
                 || !WriteFile(GetStdHandle(STD_ERROR_HANDLE), "\r\n", sizeof("\r\n") - 1, &cbWritten, NULL)
                 )
-                *(void **)i = NULL;
+                *(void **)(size_t)i = NULL;
             ExitProcess(127);
         }
     }
